backend/requirements.txt
Django==4.1.2
django-filter==23.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
pillow==10.2.0
python-dotenv==1.0.0